+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 [comment]: <> (When bumping [pc:VERSION_LATEST_RELEASE] create a new entry below)
 ### Unreleased version
 - Add support to list indexes
+- Add support to configure index
 
 ### v0.6.0
 - Add async stub for data plane operations